error handling crud

+++ src/routes/deletePokemon.js
@@ -2,14 +2,25 @@ const { Pokemon } = require("../db/sequelize");
 
 module.exports = (app) => {
   app.delete("/api/pokemons/:id", (req, res) => {
-    Pokemon.findByPk(req.params.id).then((pokemon) => {
-      const pokemonDeleted = pokemon;
-      Pokemon.destroy({
-        where: { id: pokemon.id },
-      }).then((_) => {
-        const message = `Le pokémon avec l'identifiant numéro : ${pokemonDeleted.id} à bien été supprimé.`;
-        res.json({ message, data: pokemonDeleted });
+    Pokemon.findByPk(req.params.id)
+      .then((pokemon) => {
+        if (pokemon === null) {
+          const message =
+            "Le pokémon demandé n'éxiste pas. Réessayez avec un autre identifiant.";
+          return res.status(404).json({ message });
+        }
+        const pokemonDeleted = pokemon;
+        return Pokemon.destroy({
+          where: { id: pokemon.id },
+        }).then((_) => {
+          const message = `Le pokémon avec l'identifiant numéro : ${pokemonDeleted.id} à bien été supprimé.`;
+          res.json({ message, data: pokemonDeleted });
+        });
+      })
+      .catch((error) => {
+        const message =
+          "Les pokémon n'a pas pu être modifié. Réessayez dans quelques instants.";
+        res.status(500).json({ message, data: error });
       });
-    });
   });
 };
